refactor(invite): decouple global session state management side-effects from useValidateInviteQuery hook
