+++ apps/meteor/client/views/invite/hooks/useValidateInviteQuery.ts
@@ -1,13 +1,10 @@
-import { useEndpoint, useSessionDispatch, useSetting, useToastMessageDispatch } from '@rocket.chat/ui-contexts';
+import { useEndpoint, useToastMessageDispatch } from '@rocket.chat/ui-contexts';
 import { useQuery } from '@tanstack/react-query';
 import { useTranslation } from 'react-i18next';
 
 export const useValidateInviteQuery = (token: string | undefined) => {
 	const { t } = useTranslation();
 
-	const registrationForm = useSetting('Accounts_RegistrationForm');
-
-	const setLoginDefaultState = useSessionDispatch('loginDefaultState');
 	const dispatchToastMessage = useToastMessageDispatch();
 
 	const handleValidateInviteToken = useEndpoint('POST', '/v1/validateInviteToken');
@@ -21,17 +18,6 @@ export const useValidateInviteQuery = (token: string | undefined) => {
 
 			try {
 				const { valid } = await handleValidateInviteToken({ token });
-
-				// FIXME: decouple this state management from the query
-				if (!token) {
-					return;
-				}
-
-				if (registrationForm !== 'Disabled') {
-					setLoginDefaultState('invite-register');
-				} else {
-					setLoginDefaultState('login');
-				}
 
 				if (!valid) {
 					throw new Error('Invalid token');
